+++ texas-hearing-institute/pages/ComeLater.tsx
@@ -1,9 +1,20 @@
+import { ScrollView } from 'react-native';
 import Heading from '../components/Heading';
+import Subheading from '../components/SubHeading';
 
 
 const Comelater: React.FC = () => {
+  const handleSubmission = (buttonStates: boolean[]) => {
+    console.log('Button states:', buttonStates);
+  };
+
+  const buttonLabels = ['Button 1', 'Button 2', 'Button 3'];
   return (
-    <Heading title={"Coming Soon."}></Heading>
+    <ScrollView>
+    <Heading title={"Initial Consonants"}></Heading>
+    <Subheading title={"Select a consonant and blahblah to blahblah practise speaking"}></Subheading>
+    </ScrollView>
+   
   )
 }
